+++ client/src/pages/Favories.tsx
@@ -1,6 +1,6 @@
 import "./Favories.css";
-import Header from "../components/header";
-import Nav from "../components/nav";
+import Header from "../components/Header/Header";
+import Nav from "../components/nav/Navbar";
 
 function Fav() {
   return (
